modified some files
