+++ frontend/src/components/Navbar.jsx
@@ -22,7 +22,7 @@ export default function Navbar() {
     <nav className="navbar navbar-expand-lg bg-body-tertiary">
       <div className="container-fluid">
         <NavLink className="navbar-brand" to="/">
-          My Blog
+          BlogSphere
         </NavLink>
         <button
           className="navbar-toggler"
