+++ frontend/src/pages/Blog.jsx
@@ -5,33 +5,39 @@ import Navbar from "../components/Navbar";
 import BlogCard from "../components/BlogCard";
 
 export default function Blog() {
-  function readBlog() {
-    const utterance = new SpeechSynthesisUtterance();
-    utterance.text = `In the 1990s, Sony was preparing to launch a household robot. Their
-    technology was impressive but imperfect: the robot could mishear
-    commands or simply get them wrong. So, instead of launching the
-    product as a robot (think block-y design, 3CPO-like interface,
-    computer-like voice), the team launched it as a dog.
+  // function readBlog() {
+  //   const utterance = new SpeechSynthesisUtterance();
+  //   utterance.text = `In the 1990s, Sony was preparing to launch a household robot. Their
+  //   technology was impressive but imperfect: the robot could mishear
+  //   commands or simply get them wrong. So, instead of launching the
+  //   product as a robot (think block-y design, 3CPO-like interface,
+  //   computer-like voice), the team launched it as a dog.
 
-    Lorem ipsum dolor sit amet consectetur adipisicing elit. Iste
-    distinctio odit tempora? Nihil earum dolorem, aperiam vitae debitis,
-    cum laudantium quaerat ipsa praesentium dolore omnis atque
-    aspernatur alias sint aliquid ipsum! Corporis quidem nemo maxime.
+  //   Lorem ipsum dolor sit amet consectetur adipisicing elit. Iste
+  //   distinctio odit tempora? Nihil earum dolorem, aperiam vitae debitis,
+  //   cum laudantium quaerat ipsa praesentium dolore omnis atque
+  //   aspernatur alias sint aliquid ipsum! Corporis quidem nemo maxime.
 
-    Lorem ipsum dolor sit amet consectetur adipisicing elit. Provident,
-    laboriosam aperiam velit in maiores fugiat iusto excepturi
-    blanditiis? Reiciendis tenetur at ea eaque! Enim sequi quod
-    deleniti, soluta excepturi doloribus, blanditiis molestias, officiis
-    sapiente dolorem unde labore ducimus id! Ad libero totam beatae?
-    Itaque delectus repellendus laudantium, nemo excepturi numquam
-    ratione temporibus vero cumque veniam velit a blanditiis sunt,
-    inventore veritatis commodi nobis fugit. Totam quo illum enim
-    impedit! Atque, sapiente doloribus ratione, omnis exercitationem
-    odit libero aspernatur totam possimus deserunt eaque! Ut,
-    repellendus laborum.
-    `;
-    speechSynthesis.speak(utterance);
-  }
+  //   Lorem ipsum dolor sit amet consectetur adipisicing elit. Provident,
+  //   laboriosam aperiam velit in maiores fugiat iusto excepturi
+  //   blanditiis? Reiciendis tenetur at ea eaque! Enim sequi quod
+  //   deleniti, soluta excepturi doloribus, blanditiis molestias, officiis
+  //   sapiente dolorem unde labore ducimus id! Ad libero totam beatae?
+  //   Itaque delectus repellendus laudantium, nemo excepturi numquam
+  //   ratione temporibus vero cumque veniam velit a blanditiis sunt,
+  //   inventore veritatis commodi nobis fugit. Totam quo illum enim
+  //   impedit! Atque, sapiente doloribus ratione, omnis exercitationem
+  //   odit libero aspernatur totam possimus deserunt eaque! Ut,
+  //   repellendus laborum.
+  //   `;
+  //   speechSynthesis.speak(utterance);
+  // }
+
+  // useEffect(() => {
+  //   return () => {
+  //     speechSynthesis.cancel();
+  //   };
+  // }, []);
 
   return (
     <>
@@ -74,7 +80,7 @@ export default function Blog() {
         <hr />
 
         {/* actions buttons for post */}
-        <BlogPostButtons listenOnclick={readBlog} />
+        <BlogPostButtons />
 
         <hr />
